requirements.txt
annotated-types==0.6.0
anyio==4.2.0
boto3==1.34.31
botocore==1.34.31
click==8.1.7
colorama==0.4.6
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
jmespath==1.0.1
mangum==0.17.0
numpy==1.26.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
pydantic==2.6.0
pydantic_core==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.1
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
